+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [1.0.1] - 2026-05-14
 ### Added
 - Add docker images
+- Add connection with System env and properties
 
 ### Fixed